# plugin/openapi.yaml
openapi: 3.1.0
info:
  title: Gnosis RAG API
  description: API for querying and analyzing Obsidian vaults using hybrid RAG
  version: 1.0.1
servers:
  - url: https://gnosis-rag-api.onrender.com
    description: Cloudflare Tunnel URL
paths:
...
  /dropbox/sync:
    post:
      summary: Dropbox Vault Sync
      description: |
        Performs bidirectional synchronization between Dropbox and the local Obsidian vault.
        Downloads new/updated files from Dropbox, uploads new/updated files from local vault, and resolves conflicts.
      operationId: sync_dropbox_vault_post
      responses:
        "200":
          description: Successful Dropbox Sync
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Dropbox vault synchronized successfully.
                  files_downloaded:
                    type: integer
                    example: 2
                  files_uploaded:
                    type: integer
                    example: 1
                  files_conflicted:
                    type: integer
                    example: 0
                  files_skipped:
                    type: integer
                    example: 5
                  errors:
                    type: array
                    items:
                      type: string
                  timestamp:
                    type: string
                    format: date-time
        "400":
          description: Dropbox Sync Disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /dropbox/status:
    get:
      summary: Dropbox Sync Status
      description: Returns the current Dropbox sync status and configuration.
      operationId: get_dropbox_status
      responses:
        "200":
          description: Dropbox Sync Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
                  is_syncing:
                    type: boolean
                  last_sync_time:
                    type: string
                    format: date-time
                  sync_interval_minutes:
                    type: integer
                  local_vault_path:
                    type: string
                  dropbox_connected:
                    type: boolean
                  timestamp:
                    type: string
                    format: date-time
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
components:
  schemas:
...
    ErrorDetail:
      title: ErrorDetail
      type: object
      properties:
        detail:
          type: string
...